mcp-server/requirements.txt
Flask==2.3.0
Werkzeug==2.3.0
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0